Fixing Win10 building errors and warnings
Former-commit-id: 4c56d98acdab369a45f57d27e02c5bef323f584b

+++ Sources/Internal/Render/Highlevel/StaticOcclusion.cpp
@@ -213,7 +213,7 @@ namespace DAVA
         
         AABBox3 cellBox = GetCellBox(cellX, cellY, cellZ);
         Vector3 stepSize = cellBox.GetSize();
-        stepSize /= stepCount;
+        stepSize /= float32(stepCount);
         
         
         Vector3 directions[6] =
